Update README.md to clarify dashboard options and add new Streamlit dashboard script
