+++ dashboard/README.md
@@ -1,6 +1,6 @@
 # Dashboard
 
-Some dashboards live here.
+Some simple local dashboard options live here.
 
-- [streamlit.py](streamlit.py): [Streamlit](https://streamlit.io/) dashboard for the project.
-- [fasthtml.py](fasthtml.py): [FastHTML](https://fastht.ml/) dashboard for the project.
+- [`streamlit-dashboard.py`](streamlit-dashboard.py): [Streamlit](https://streamlit.io/) dashboard for the project.
+- [`fasthtml-dashboard.py`](fasthtml-dashboard.py): [FastHTML](https://fastht.ml/) dashboard for the project (WIP).
